// src/constant/message.interface.ts
export enum Message {
  SUCCESS = "Success",
  CREATED = "Created",
  UPDATED = "Updated",
  DELETED = "Deleted Successfull",
  BAD_REQUEST = "Bad Request",
  UNAUTHORIZED = "Unauthorized",
  SESSION_EXPIRED = "Session Expired",
  INTERNAL_SERVER_ERROR = "Internal Server Error",
  NOT_FOUND = "NOT Found",
  FORBIDDEN = "Forbidden",
  PAGE_NOT_FOUND = "page not found",
  FOUND = "Found",
}